requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
python-dotenv==1.0.0
httpx==0.26.0
requests-oauthlib==1.3.1
